agents/twitter-scraper/requirements.txt
openai-agents>=0.1.0
playwright>=1.40.0
agentql>=0.5.0
pydantic>=2.0
python-dotenv>=1.0.0
